@@ -47,15 +47,19 @@
     "@types/react": "^16.4.6",
     "@types/redux-logger": "^3.0.7",
     "@types/webpack-env": "^1.13.6",
+    "@typescript-eslint/eslint-plugin": "^2.3.1",
+    "@typescript-eslint/parser": "^2.3.1",
     "babel-eslint": "^8.2.3",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-jsx-stylesheet": "^0.6.5",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-env": "^1.6.1",
-    "eslint": "^4.19.1",
-    "eslint-config-taro": "1.2.14",
+    "eslint": "^6.4.0",
+    "eslint-config-prettier": "^6.3.0",
+    "eslint-config-taro": "^1.3.19",
     "eslint-plugin-import": "^2.12.0",
+    "eslint-plugin-prettier": "^3.1.1",
     "eslint-plugin-react": "^7.8.2",
     "eslint-plugin-taro": "1.2.14",
     "eslint-plugin-typescript": "^0.12.0",
