@@ -1,15 +1,14 @@
 export class User {
   request: any;
-  RESTful: string;
+  Prefix: string;
 
   constructor(request: any) {
     this.request = request;
-    this.RESTful = 'api/user';
+    this.Prefix = 'api/user';
   }
 
-  login = (data: { code: string }) => {
-    return this.request.post(`${this.RESTful}/login`, {
+  login = (data: { code: string }) =>
+    this.request.post(`${this.Prefix}/login`, {
       data
     }) as Promise<any>;
-  };
 }
