@@ -2,7 +2,6 @@ import { catchError, concat, mergeMap, switchMap } from 'rxjs/operators';
 import { combineEpics, ofType } from 'redux-observable';
 import { of, from } from 'rxjs';
 import { IEpic } from '@common/helpers';
-
 import * as errorActions from '@models/global/error/actions';
 import * as actions from './actions';
 
@@ -14,11 +13,9 @@ const login: IEpic<any> = (action$, state$, { api }) =>
         mergeMap(response => of(actions.login.done(response)))
       )
     ),
-    catchError((e, source$) => {
-      return of(actions.login.error(), errorActions.capture(e)).pipe(
-        concat(source$)
-      );
-    })
+    catchError((e, source$) =>
+      of(actions.login.error(), errorActions.capture(e)).pipe(concat(source$))
+    )
   );
 
 export const epic = combineEpics(login);
