@@ -1,6 +1,6 @@
 import Taro, { Component } from '@tarojs/taro';
 import { View, Text, Swiper, SwiperItem } from '@tarojs/components';
-import Iconfont from '@components/iconfont/index';
+import { Iconfont } from '@components';
 
 import './index.less';
 
