@@ -7,7 +7,9 @@ export default class Quiz extends Component {
     return (
       <View className='quiz'>
         <Text className='title'>不确定如何开始</Text>
-        <Navigator url='' className='navigator'>Take the quiz</Navigator>
+        <Navigator url='' className='navigator'>
+          Take the quiz
+        </Navigator>
       </View>
     );
   }
